Wolrd Object Entity - Reorg
