+++ Source/ACE/ACE.csproj
@@ -97,14 +97,14 @@
     <Compile Include="Command\Handlers\HelpCommands.cs" />
     <Compile Include="Entity\Adjacency.cs" />
     <Compile Include="Entity\BroadcastAction.cs" />
-    <Compile Include="Entity\CollidableObject.cs" />
-    <Compile Include="Entity\Container.cs" />
-    <Compile Include="Entity\Creature.cs" />
-    <Compile Include="Entity\Door.cs" />
+    <Compile Include="Entity\WorldObjects\CollidableObject.cs" />
+    <Compile Include="Entity\WorldObjects\Container.cs" />
+    <Compile Include="Entity\WorldObjects\Containers\Creature.cs" />
+    <Compile Include="Entity\WorldObjects\UsableObjects\Door.cs" />
     <Compile Include="Entity\Events\BroadcastEventArgs.cs" />
     <Compile Include="Entity\Events\ChatMessageArgs.cs" />
-    <Compile Include="Entity\Lifestone.cs" />
-    <Compile Include="Entity\Monster.cs" />
+    <Compile Include="Entity\WorldObjects\UsableObjects\Lifestone.cs" />
+    <Compile Include="Entity\WorldObjects\Containers\Creatures\Monster.cs" />
     <Compile Include="Entity\MovementData.cs" />
     <Compile Include="Entity\Events\SpellCastArgs.cs" />
     <Compile Include="Entity\Landblock.cs" />
@@ -115,10 +115,10 @@
     <Compile Include="Entity\ModelTexture.cs" />
     <Compile Include="Entity\ModelData.cs" />
     <Compile Include="Entity\PhysicsData.cs" />
-    <Compile Include="Entity\Portal.cs" />
-    <Compile Include="Entity\SpellLikeEffect.cs" />
-    <Compile Include="Entity\SummonedPortal.cs" />
-    <Compile Include="Entity\UsableObject.cs" />
+    <Compile Include="Entity\WorldObjects\CollidableObjects\Portal.cs" />
+    <Compile Include="Entity\WorldObjects\SpellLikeEffect.cs" />
+    <Compile Include="Entity\WorldObjects\UsableObjects\SummonedPortal.cs" />
+    <Compile Include="Entity\WorldObjects\UsableObject.cs" />
     <Compile Include="Factories\CommonObjectFactory.cs" />
     <Compile Include="Factories\CorpseObjectFactory.cs" />
     <Compile Include="Factories\GenericObjectFactory.cs" />
@@ -199,8 +199,8 @@
     <Compile Include="Network\GameAction\GameActionType.cs" />
     <Compile Include="Network\Enum\PhysicsDescriptionFlag.cs" />
     <Compile Include="Network\Enum\WeenieHeaderFlags.cs" />
-    <Compile Include="Entity\Player.cs" />
-    <Compile Include="Entity\WorldObject.cs" />
+    <Compile Include="Entity\WorldObjects\Containers\Creatures\Player.cs" />
+    <Compile Include="Entity\WorldObjects\WorldObject.cs" />
     <Compile Include="Managers\WorldManager.cs" />
     <Compile Include="Network\Enum\CharacterError.cs" />
     <Compile Include="Network\GameAction\Actions\GameActionMoveToState.cs" />
