+++ Source/ACE/Entity/WorldObjects/CollidableObject.cs
