+++ Source/ACE/Entity/WorldObjects/CollidableObjects/Portal.cs
