+++ Source/ACE/Entity/WorldObjects/Container.cs
