+++ Source/ACE/Entity/WorldObjects/Containers/Creature.cs
