+++ Source/ACE/Entity/WorldObjects/SpellLikeEffect.cs
