+++ Source/ACE/Entity/WorldObjects/UsableObject.cs
