+++ Source/ACE/Entity/WorldObjects/UsableObjects/Lifestone.cs
