+++ Source/ACE/Entity/WorldObjects/WorldObject.cs
